Cambio en el service y test para operar con constantes. FALLA TEST CC PROVEEDORES.

+++ src/test/java/sic/modelo/dto/NotaDebitoDTO.java
@@ -1,6 +1,7 @@
 package sic.modelo.dto;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -12,7 +13,7 @@ import sic.modelo.RenglonNotaDebito;
 public class NotaDebitoDTO extends NotaDTO implements Serializable {
     
     private List<RenglonNotaDebito> renglonesNotaDebito;
-    private double montoNoGravado;
+    private BigDecimal montoNoGravado;
     private Recibo recibo;
     
 }
